requirements.txt
flask==3.1.0
anthropic==0.40.0
qrcode[pil]==8.0
PyMuPDF==1.25.1
Pillow==11.1.0
gunicorn==23.0.0